add auth

+++ webpack.mix.js
@@ -1,6 +1,8 @@
-const mix = require("laravel-mix");
-require("laravel-mix-tailwind");
-require("laravel-mix-purgecss");
+const mix = require('laravel-mix');
+
+require('laravel-mix-tailwind');
+require('laravel-mix-purgecss');
+
 /*
  |--------------------------------------------------------------------------
  | Mix Asset Management
@@ -12,7 +14,12 @@ require("laravel-mix-purgecss");
  |
  */
 
-mix.js("resources/js/app.js", "public/js")
-    .sass("resources/sass/app.scss", "public/css")
-    .tailwind()
-    .purgeCss();
+mix.js('resources/js/app.js', 'public/js')
+   .postCss('resources/css/app.css', 'public/css')
+   .tailwind('./tailwind.config.js');
+
+if (mix.inProduction()) {
+  mix
+   .version()
+   .purgeCss();
+}
